site._loginstatus - login process and caching
site._loginstatus was very unreliable due to site.login() not setting it
1. if the server considered the bot to be
   logged in without authentication, and
2. exceptions occurred during authentication

When logged in as a user, but unable to login as a sysop, the site
object should return to being logged in as a user.  LoginManager
throws exceptions, such as NoUsername, which need to be caught and
re-raised after site._loginstatus is updated.

While LoginManager is attempting to login, do not cache anything.

In cache key generation, dont instantiate a User() as the site object
may not be usable yet.

Bug: T100965
Change-Id: Ia2d37875adeabbebfc51da3a20a77b32780ccadd