Allow github clones to easily run travis builds
The wiki user 'Pywikibot-test' only works on the 'wikimedia'
travis-ci account.  This prevents unmodified clones of the
codebase from running tests on travis-ci.

The tests now run according to whatever is in user-config.py,
skipping tests which cant be run.

The default .travis.yml now detects if it is run by the 'wikimedia'
account, and detects if the travis user has a password set.

If a username or password can't be found, the username and password
are not set up, which disables any tests which require a user.

Change-Id: I0f22caaf8b5cd314a196a5e3cc25ae830231ccb5