Require flickrapi 1.4.5 for Python 2
flickrapi 1.4.4 installs a logger, which breaks script tests.
flickrapi 1.4.5 and flickrapi 2.x do not have this problem.
Require 1.4.5 to avoid this problem, and to remain on a stable
version which is easy to install on unix systems.

This issue is not relevant to Python 3, as only flickrapi 2.x
will work on Python 3, and setuptools automatically selects it.

Bug: 71454
Change-Id: I58b732a3dae69797fe9bc27067a666c091be2e93