api.py: buffer data in QueryGenerator
QueryGenerator yields all items in resultdata, but there is no
guarantee that all data for an item arrived in one response.
Some data will appear in the following response, etc.
(see commit 3d2ca97aaac17a28e177fa89fbdd3364e7f53c0c)

This patch buffers results until all data are for an item are fetched.
This is based on the fact that the API, when query-continuing, keeps on
repeating the same pages until all requested data are fetched.

Change-Id: Iccb3a96b0248fdab0650edfda23d05ecec0dadbd