Page.namespace as property returning Namespace
After the introduction of Namespace in 0f2226c6, several improvements
have been made to better support the old functionality where
Page.namespace() was an int.  However, nothing can help the
literal conversion to str using str(Page.namespace()), as the Namespace
object str representation needs to be the name and not the int.

This change restores the original behaviour of Page.namespace() returning
an actual int, and allows Page.namespace to be used as a property to
obtain the Namespace object.

Mark some reflinks tests as expectedFailure, as they only succeeded
while the Page.namespace() result was a Namespace object.

Change-Id: Idb963066a28959aaec93ec10335b99eccac49a58