Allow unexpectedSuccess and expectedFailure
Python 3.4, and unittest2 0.8, error on unexpected successes
and warn on expected failure, if those conditions are not handled.
One way to handle them is to add hooks on the TestCase subclass,
_addUnexpectedSuccess and _addExpectedFailure.  As these are used
in pywikibot tests to capture current behaviour, report to the
UI 'unexpected success' and 'expected failure', and mark as success.

Change-Id: I8468cb4507414d55f38ec1f73beda311029f4f44