Fix TestDryPageGenerator so it is not special
api_tests.TestDryPageGenerator pushed emulated query response data
directly into QueryGenerator as a dict, which also allowed the method
test_limits to reuse the generator.  This breaks the definition of a
'generator', and relied on QueryGenerator not fetching data if 'data'
already existed.

Fixed the tests so that the insertion of the mock data doesnt depend
on those quirks of QueryGenerator which should be removed.

Change-Id: Ic4e92dc6eefe578fa12cab2307e45e7e85f96d40