Simplify QueryGenerator.__iter__ flow control
Rearrange the logic to use less nested branching, and
destroy the data before exiting the iteration.

Removes 'None' as a valid value for QueryGenerator.limit;
it was not able to be set using set_maximum_items().

Adds some iterator state information to help with debugging
and test validation, and dont leave data in an instance variable.

Bug: 64489

Change-Id: Ie3a9e3576a0174703bcccf0d77ed42d56da0d09f