Load mw and non-mw sites without using family file
This method loads the site object of type APISite for mw sites and
NonMWAPISite for non-mw sites. API path of the mw site in consideration is
determined and using the API path, the site object is created and returned.
All this is done without creating or using family files. Unittest test for
the method is also written and the test has been added to tox.

Bug: T88601
Change-Id: Ifb8c5930c21e4a90b08ef958329e22a0b97a57ae