Fix and test interwiki_graph
interwiki_graph used compat method Family.get_address, which doesnt
exist in core and caused an exception when creating a graph.

The module also relied on a 'Subject' data structure which was only
defined within interwiki.py.  A new simple Subject class has been
added to the interwiki_graph module, and the more advanced class
in interwiki.py is now a subclass.

Added basic dry tests for this module.

Change-Id: If1f41c27a7fec012e9a7c4ce7cb13bb0ecd11c2e