Change i18n.input test lang from arz to nn
The test test_pagegen_i18n_input requires the
test site code is a known locale, and doesnt have
a translation for the i18n message, causing the
i18n to fallback to another language with a translation.

The current test code 'arz' is not recognised by locale.
Change the code to 'nn', which falls back to 'nb',
and update the test to confirm that the translation
for 'nb' is shown when the locale is 'nn' using LC_ALL
environment variable and config.userinterface_lang.

Also expose a list of modified configuration variables
so that the test class can fail if config.userinterface_lang
has been set to any value other than the default None.

Bug: T128991
Change-Id: I6c54f902285101b9c13af4e6b2b75d4237cb3f3b