i18n test: use an.wp instead of arz.wp
The test test_pagegen_i18n_input requires the
test site code is a known locale, and doesnt have
a translation for the i18n message, causing the
i18n to fallback to English.

arz is not a recognised locale.
an is configured in i18n.py to fall back to es.

Change-Id: I6c54f902285101b9c13af4e6b2b75d4237cb3f3b