Dynamically determine site for i18n.input test
The test test_pagegen_i18n_input requires the
test site code is a known locale, and doesnt have
a translation for the i18n message, causing the
i18n to fallback to English.

arz is not recognised by locale.
Use a dynamic algorithm, which currently selects
af, which defaults to nl.

Change-Id: I6c54f902285101b9c13af4e6b2b75d4237cb3f3b