APISite fallback to BaseSite for old versions
When the APISite detects an unsupported version for
a method, usually using the need_version decorator,
it raises NotImplementedError.  If possible, this
should cause the same method in BaseSite to be used
as the fallback.

Implement this for redirectRegex which uses getmagicwords,
a feature available in 1.14+, by removing the method in
the subclass.

Fix an overly broad APIError exception caught in APISite.version,
add replace hard-coded family class fallback with fallback to
BaseSite which will then fallback to the Family class.

Also remove very selective set of methods which are reimplemented
as a simple method that raises NotImplementedError.  Only doing
this for some methods implies that the other inherited methods are
supported, whereas that is now tracked using @deprecated decorators.

Change-Id: If25181bf488afc0ef87071d1e6cc1ac10767461c