APISite.editpage should not reload the Page
APISite.editpage reloads the revisions after success in order
to ensure that the Page is up to date, as the actual text of the
page may be different from the submitted text due to merges
performed by the server to resolve edit conflicts.

APISite and api achieved this by modifying private variables
of BasePage.

By exposing a public setter and deleter for latest_revision_id,
APISite and api can use this property to inform BasePage of
when its caches are invalid, so it can clear the caches so
subsequent fetches retrieve the actual text of the saved revision
when it is needed.

Normalise WikibasePage to using attribute _revid instead of lastrevid,
deprecating the latter.

Also add missing text deleter to ProofreadPage.

Change-Id: I55a49d115a8f207216f24f9acbfa64d87da88578