Decommission support for Python 2.7.0 and 2.7.1
Pywikibot had problems with unicode on Python 2, resulting
in Python issue #10254, which was resolved in Python 2.7.2
and backported to Python 2.6.6.

To workaround this issue, or provide a warning, would require
checking titles in the Link constructor for the change that
the page title could cause failure.

It is unlikely that supporting Python 2.7.0 and 2.7.1 is still
necessary.

Moves the unicode test from the version script to the test suite.

Bug: T102461
Change-Id: If4a8e8eed682b837dd468b02a5d98b01ebf37584