Python issue #10254
Pywikibot had problems with NFC normalizing unicode,
resulting in Python issue #10254, which was resolved
in Python 2.7.2 and backported to Python 2.6.7.

To workaround this issue, or provide a warning, requires
checking titles in the Link constructor for titles that might
cause the failure.

It is unlikely that supporting Python 2.7.0 and 2.7.1 is still
necessary, so decommission support for Python 2.7.0 and 2.7.1.

For Python 2.6.6, the Link constructor raises UnicodeError for
titles containing combining characters, which will allow normal
usage for many languages, but will fail predictably for page
titles which could encounter this error.

Moves the unicode test from the version script to the test suite,
as the only supported platform with this bug is Python 2.6.6.

Bug: T102461
Change-Id: If4a8e8eed682b837dd468b02a5d98b01ebf37584