Fix Family.from_url support for generated families
The generator creates family classes which require explicit
code parameters to each method.  This failed for multi-code
generated families, as the domain name was expected to be in
regex group 0, however the constructed regex places the domain in
one of the regex group, but not necessarily the first regex group.

Also better optimise from_url for family classes which have methods
that do not accept None for the code.

Bug: T85658
Change-Id: I266b39d359d69b7fdd940fa337c14c3782c66432