Sync requirements.txt with setup.py
Flag pYsearch as an external unverified package.

Use git repo for pydot fork which supports Python 3.

Use the environment markers introduced in pip 6.0
to prevent installation of unnecessary or broken packages
based on the Python version.

mwlib is not included, as it is incompatible with
other dependencies.

Also create dev-requirements.txt for dev/test dependencies,
and use this in .travis.yml and .appveyor.yml
pywin32 and pywinauto are not included as they do not
have any effect on the Appveyor builds.

Change-Id: I6c19152c544a38c21c79d9311fbfbd211c5a4b43