[IMPROV] UI: Always use color stack
Previously only the colored Windows UI used a color stack because of the
transliteration. This is implementing the stack in general and each UI only
implements a different way to specify the color.

It is also removing the ambiguity between using the default color as the actual
default color and using the last color from the stack by adding a new color
`previous`.

Change-Id: Iecbfe7b633b3cb5dbe1772cbccb03641637a6612