Remove implicit use of `pywikibot`
`pywikibot` was being used without an 'import pywikibot' and
this wasnt a syntax error as 'import pywikibot.tools' or similar
implicitly imports `pywikibot`.
However this import may not always occur if the sub-module has
already been loaded, such as if the package import fails but
some sub-modules have been imported.

Bug: T113161
Change-Id: Id01888d3b9d847927d4adce236153ac4b91a0c72