[FEAT] site: Dynamically determine Wikibase repo
Instead of statically defining the Wikibase repository in the family, it can
query the API to determine the repository without user input. This also changes
the implementation in `Family` to use the site's implementation.

It also makes it necessary to specify the data repository used on dry tests as
they cannot query the API to get the data repository.

Bug: T85331
Change-Id: I5568d848a369c8af36a40c193e2409071e23a5fb