Detect connection refused for TLSv1.2 only sites
Ubuntu precise prior to May 2015 did not auto-negotiate TLSv1.2.
It needed to be specifically requested.  This is the current
situation with the Travis precise images, with InterWikiMap entry
'http://www.proofwiki.org/wiki/$1'
and is further complicated because that redirects to
'http://proofwiki.org/wiki/$1'

As this occurs only on an old unpatched Ubuntu release, and most users
will have automatically patched their system, and very few sites are
now TLSv1.2 only, this change only implements:
1. detection on Python 3.4+ for whether the connection succeeds if
   TLSv1.2 is explicitly requested, as Python 3.4+ has ssl.PROTOCOL_TLSv1_2
2. a simple fallback for lower versions of Python when connecting to
   wikis known to have this problem.
In both cases the uninformative ConnectionError is replaced with a
FatalServerError with a descriptive message.
The solution is to patch the machine.

Bug: T134647
Change-Id: Icb374767daf963ee31eab4b16fa8a4f86b57afd4